modified js files
