+++ src/components/Calculator.js
@@ -1,19 +1,31 @@
 import React from 'react';
 import './Calculator.css';
-import PropTypes from 'prop-types';
+import calculate from '../logic/calculate';
 
 class Calculator extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {};
+    this.state = { total: null, next: null, operation: null };
   }
 
   render() {
-    const { updateDetails, value } = this.props;
+    const updateDetails = (e) => {
+      const newObj = calculate(this.state, e.target.textContent);
+      this.setState(newObj);
+    };
+
+    const { next, operation, total } = this.state;
+    const op = operation === '%' ? 'mod' : operation;
+    let result = '';
+    if (total) {
+      result = `${total} ${op || ''} ${next || ''}`;
+    } else if (next) {
+      result = `${next} ${op || ''}`;
+    }
     return (
       <div className="container">
         <div className="screen">
-          {value || 0}
+          {result || 0}
         </div>
         <div className="buttons">
           <button onClick={updateDetails} type="button">AC</button>
@@ -23,11 +35,11 @@ class Calculator extends React.Component {
           <button onClick={updateDetails} type="button">7</button>
           <button onClick={updateDetails} type="button">8</button>
           <button onClick={updateDetails} type="button">9</button>
-          <button onClick={updateDetails} className="orange" type="button">&times;</button>
+          <button onClick={updateDetails} className="orange" type="button">x</button>
           <button onClick={updateDetails} type="button">4</button>
           <button onClick={updateDetails} type="button">5</button>
           <button onClick={updateDetails} type="button">6</button>
-          <button onClick={updateDetails} className="orange" type="button">&minus;</button>
+          <button onClick={updateDetails} className="orange" type="button">-</button>
           <button onClick={updateDetails} type="button">1</button>
           <button onClick={updateDetails} type="button">2</button>
           <button onClick={updateDetails} type="button">3</button>
@@ -40,10 +52,5 @@ class Calculator extends React.Component {
     );
   }
 }
-
-Calculator.propTypes = {
-  updateDetails: PropTypes.func.isRequired,
-  value: PropTypes.string.isRequired,
-};
 
 export default Calculator;
